Remove 'baseurl' logic
We have had problems getting the "About" page working properly. The
link highlight logic for the currently active page also was not working.
The reason for both problems seems to be that we use logic involving the
'baseurl' variable. This may or may not be connected to a change that
required us to remove the 'relative_permalinks' setting from
_config.yml.
In any case, this change removes this logic and just hard codes the
path. Note that we had been doing that all along for the recently
introduced logos inside the style sheet.